Hospital kann mit vollen Leben genutzt werden Loanshark Grundgerüst steht, die zwei hauptmethoden müssen nun implementiert werden

+++ src/main/java/inputHandler/BuyHandler.java
@@ -65,7 +65,7 @@ public class BuyHandler {
 
     private int maxBeansAbleToBuy(int pickedBeanNumber) {
         int result;
-        result = (int) (state.getCash() / state.getPriceOf(pickedBeanNumber));
+        result = state.getCash() / state.getPriceOf(pickedBeanNumber);
         return result;
     }
